refactor: introduce new interface for test cases
To improve the readabiliy and maintainability the Arguments where extracted into an interface and better documented. The separate test cases got their own classes and are used in the AsymmetricNetowrkPartitionIT test
